refact router.php - create routerFront.php
