+++ config/RouterBack.php
@@ -2,203 +2,8 @@
 
 namespace config;
 
-class Router 
+class RouterBack
 { 
-    public function runFrontPage($get)
-    {
-        if ($get == 'home') {
-
-            $frontController = new \controller\FrontController;
-
-            if (!isset($_POST['submitMessage']) && isset($_SESSION['msg'])) {
-                $msg = $_SESSION['msg'];
-                $frontController -> homePage($msg);
-                unset($_SESSION['msg']);
-                return;
-            } else {
-                $frontController -> homePage();
-            }   
-
-            if (!empty($_POST['inputName']) 
-                && !empty($_POST['inputFirstName']) 
-                && !empty($_POST['inputEmail']) 
-                && !empty($_POST['inputMessage'])
-            ) {
-                $form = array('inputName' => $_POST['inputName'], 
-                        'inputFirstName' => $_POST['inputFirstName'], 
-                        'inputEmail' => $_POST['inputEmail'], 
-                        'inputMessage' => filter_input(
-                            INPUT_POST, 'inputMessage', 
-                            FILTER_SANITIZE_SPECIAL_CHARS
-                        )
-                );
-                $msg = $this -> runSendMessage($form); 
-
-            } 
-            $msg = EMPTY_FIELDS;
-            $_SESSION['msg'] = $msg;
-
-            header('Location: index.php?p=home#signup'); 
-            exit();
-
-            
-        } elseif ($get == 'listposts') {
-
-            $frontController = new \controller\FrontController;
-            $frontController -> listPostsView(); 
-
-            
-        } elseif ($get == 'post') {
-
-            if (isset($_GET['id'])) {
-                $frontController = new \controller\FrontController;
-                $backCommentController = new \controller\BackCommentController;
-
-                if (isset($_GET['c']) 
-                    && ($_GET['c']=='ok' || $_GET['c']=='moderate')
-                ) {
-                    if (isset($_GET['cid'])) {
-                    
-                        $backCommentController -> updateComment(
-                            filter_input(
-                                INPUT_GET, 'cid',
-                                FILTER_SANITIZE_SPECIAL_CHARS
-                            )
-                        );
-                    }
-                    $frontController -> postView($_GET['id']);
-
-                } elseif (isset($_POST['submitComment'])) {
-                    
-                    if (!empty($_POST['nameVisitor']) 
-                        && !empty($_POST['emailVisitor']) 
-                        && !empty($_POST['content'])
-                    ) {
-                        
-                        if (isset($_SESSION['user_id'])) {
-                            $user_id = $_SESSION['user_id'];
-                        } else {
-                            $user_id = null;
-                        }
-                        
-                        $form = array(
-                            'nameVisitor' => $_POST['nameVisitor'],
-                            'emailVisitor' => $_POST['emailVisitor'],
-                            'content' => $_POST['content'],
-                            'validComment' => 'FALSE',
-                            'user_id' => $user_id,
-                            'post_id' =>$_GET['id']
-                        );
-                        
-                        $msg = $frontController -> addNewComment($form); 
-                        
-                    } else {
-                        $msg = EMPTY_FIELDS;
-                    }
-
-                    $_SESSION['commentMsg'] = $msg;
-
-                    header('Location: index.php?p=post&id=' . $_GET['id'] . '#comments');
-                    exit();
-                
-                } elseif (isset($_POST['submitModerateComment'])) {
-
-                    if (!empty($_POST['nameVisitor']) 
-                        && !empty($_POST['emailVisitor']) 
-                        && !empty($_POST['content'])
-                    ) {
-
-
-                        $form = array(
-                            'nameVisitor' => 'Admin : ' . $_POST['nameVisitor'],
-                            'emailVisitor' => $_POST['emailVisitor'],
-                            'content' => $_POST['content'],
-                            'validComment' => 'TRUE',
-                            'user_id' => $_SESSION['user_id'],
-                            'post_id' =>$_GET['id']
-                        );
-                        
-                        $frontController -> addNewComment($form); 
-                        
-                    } else {
-                        $msg = EMPTY_FIELDS;
-                    }
-                    $_SESSION['commentMsg'] = $msg;
-                    header('Location: index.php?p=post&id=' . $_GET['id'] . '&c=ok');
-
-                } elseif (isset($_POST['publishedPost'])) {
-                    
-                        $backPostController = new \controller\BackPostController;
-                        $backPostController -> publishedPost($_GET['id']);
-
-                } else {
-                    
-                    if (isset($_SESSION['commentMsg'])) {
-                        $frontController -> postView(
-                            $_GET['id'], $_SESSION['commentMsg']
-                        );
-                        unset($_SESSION['commentMsg']);
-                    } else {
-                        $frontController -> postView($_GET['id']);
-                    }   
-                }
-
-            } else {
-                throw new \Exception(PAGE_NOT_EXIST);
-            }
-
-        } elseif ($get == 'connect') {
-            
-            $frontController = new \controller\FrontController;
-
-            if (isset($_POST['submitConnect'])) {
-                
-                if (!empty($_POST['inputPseudoConnect']) 
-                    && !empty($_POST['inputPasswordConnect'])
-                ) {
-
-                    $msgConnect = $frontController -> verifyUser(
-                        $_POST['inputPseudoConnect'], 
-                        $_POST['inputPasswordConnect']    
-                    );
-                    
-                    $_SESSION['msgConnect'] = $msgConnect;
-                    
-                    header('Location: index.php?p=connect');
-                    exit();
-
-                } else {
-
-                    $msgConnect = EMPTY_FIELDS;
-                    $frontController -> connectView($msgConnect);
-
-                }
-            } else {
-            
-                if (isset($_SESSION['msgConnect'])) {
-
-                    $frontController -> connectView($_SESSION['msgConnect']);
-                    unset($_SESSION['msgConnect']);
-
-                } else {
-                    $frontController -> connectView();
-                }
-            }    
-        } elseif ($get == 'disconnect') {
-
-                unset($_SESSION['user']);
-                header('Location: index.php');
-        }
-    }
-
-    public function runSendMessage(array $form)
-    {
-        $controller = new \controller\Controller;
-        $controller -> sendMessage($form, SUPPORT_EMAIL);
-        $msg = $controller -> msg();
-        return $msg;
-    }
-
     public function runBackPage($get)
     {
         if (isset($_SESSION['user'])) {
