+++ controller/frontController.php
@@ -27,6 +27,37 @@ class FrontController extends Controller
         );
     }
 
+    public function testInputMessage()
+    {
+        if (!empty($_POST['inputName']) 
+            && !empty($_POST['inputFirstName']) 
+            && !empty($_POST['inputEmail']) 
+            && !empty($_POST['inputMessage'])
+        ) {
+            $form = array('inputName' => $_POST['inputName'], 
+                    'inputFirstName' => $_POST['inputFirstName'], 
+                    'inputEmail' => $_POST['inputEmail'], 
+                    'inputMessage' => filter_input(
+                        INPUT_POST, 'inputMessage', 
+                        FILTER_SANITIZE_SPECIAL_CHARS
+                    )
+            );
+            return $form;
+        }
+        return EMPTY_FIELDS;
+    }
+
+    public function runSendMessage($form)
+    {
+        $controller = new \controller\Controller;
+        if ($form == EMPTY_FIELDS) {
+            return EMPTY_FIELDS;
+        }
+        $controller -> sendMessage($form, SUPPORT_EMAIL);
+        $msg = $controller -> msg();
+        return $msg;
+    }
+
     public function listPostsView()
     {
         $postManager = new \model\PostManager;
@@ -94,6 +125,53 @@ class FrontController extends Controller
         }
     }
 
+    public function validComment()
+    {
+        $backCommentController = new \controller\BackCommentController;
+
+        if (isset($_GET['cid'])) {      
+            $backCommentController -> updateComment(
+                filter_input(
+                    INPUT_GET, 'cid',
+                    FILTER_SANITIZE_SPECIAL_CHARS
+                )
+            );
+            return;
+        }
+    }
+
+    public function testInputComment($valid, $moderate=null)
+    {
+        $frontController = new \controller\FrontController;
+
+        $userId = null;
+        $userType = null;
+
+        if (isset($_SESSION['user'])) {
+            $userId = $_SESSION['user']->userId();
+        } 
+
+        if ($moderate != null) {
+            $userType = $_SESSION['user']->type() . ' : ';
+        }
+
+        if (!empty($_POST['nameVisitor']) 
+            && !empty($_POST['emailVisitor']) 
+            && !empty($_POST['content'])
+        ) {                           
+            $form = array(
+                'nameVisitor' => $userType . $_POST['nameVisitor'],
+                'emailVisitor' => $_POST['emailVisitor'],
+                'content' => $_POST['content'],
+                'validComment' => $valid,
+                'user_id' => $userId,
+                'post_id' =>$_GET['id']
+            );           
+            return $this -> addNewComment($form);              
+        } 
+        return EMPTY_FIELDS;             
+    }
+
     public function connectView($msg=null)
     {
         $this->twigInit();
@@ -105,6 +183,25 @@ class FrontController extends Controller
             )
         );
 
+    }
+
+    public function testInputConnect()
+    {
+        if (!empty($_POST['inputPseudoConnect']) 
+            && !empty($_POST['inputPasswordConnect'])
+        ) {
+            $msgConnect = $this -> verifyUser(
+                $_POST['inputPseudoConnect'], 
+                $_POST['inputPasswordConnect']    
+            );
+            return $msgConnect;
+            
+        } else {
+
+            $msgConnect = EMPTY_FIELDS;
+            $this -> connectView($msgConnect);
+            return;
+        }
     }
 
     public function verifyUser($pseudo, $password)
