+++ view/backView/validCommentView.twig
@@ -22,7 +22,7 @@
         <hr class="d-none d-lg-block mb-2 ml-0" />
         {% for comment in comments %}
             <div class="row mt-3 mb-3">
-                <div class="col-lg-1 align-self-center"><span class="badge badge-dark">{{comment.commentId}}</span></div>
+                <div class="col-lg-1 align-self-center"><span class="badge badge-dark">{{comment.id}}</span></div>
                 <div class="col-lg-1 text-truncate align-self-center">{{ comment.nameVisitor }}</div>
                 <div class="col-lg-4 align-self-center text-size-sm">{{ comment.content }}</div>
                 <div class="col-lg-2 align-self-center">{{ comment.commentDate|date("d/m/Y") }}</div>
@@ -37,7 +37,7 @@
                 </div>
                 <!--Moderate-->
                 <div class="col-lg-1 align-self-center text-center">
-                    <a href={{'index.php?p=post&id=' ~ comment.post_id ~ '&c=moderate&cid=' ~ comment.commentId ~ '#addComment'}} class="btn btn-sm text-success font-weight-bold">
+                    <a href={{'index.php?p=post&id=' ~ comment.post_id ~ '&c=moderate&cid=' ~ comment.id ~ '#addComment'}} class="btn btn-sm text-success font-weight-bold">
                         <svg width="2em" height="2em" viewBox="0 0 16 16" class="bi bi-chat-text" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
                             <path fill-rule="evenodd" d="M2.678 11.894a1 1 0 0 1 .287.801 10.97 10.97 0 0 1-.398 2c1.395-.323 2.247-.697 2.634-.893a1 1 0 0 1 .71-.074A8.06 8.06 0 0 0 8 14c3.996 0 7-2.807 7-6 0-3.192-3.004-6-7-6S1 4.808 1 8c0 1.468.617 2.83 1.678 3.894zm-.493 3.905a21.682 21.682 0 0 1-.713.129c-.2.032-.352-.176-.273-.362a9.68 9.68 0 0 0 .244-.637l.003-.01c.248-.72.45-1.548.524-2.319C.743 11.37 0 9.76 0 8c0-3.866 3.582-7 8-7s8 3.134 8 7-3.582 7-8 7a9.06 9.06 0 0 1-2.347-.306c-.52.263-1.639.742-3.468 1.105z"/>
                             <path fill-rule="evenodd" d="M4 5.5a.5.5 0 0 1 .5-.5h7a.5.5 0 0 1 0 1h-7a.5.5 0 0 1-.5-.5zM4 8a.5.5 0 0 1 .5-.5h7a.5.5 0 0 1 0 1h-7A.5.5 0 0 1 4 8zm0 2.5a.5.5 0 0 1 .5-.5h4a.5.5 0 0 1 0 1h-4a.5.5 0 0 1-.5-.5z"/>
@@ -49,7 +49,7 @@
                     <div class="row">
                         <!--Delete-->
                         <div class="col-lg-6 ">
-                            <a href={{'index.php?admin=comment&delete=' ~ comment.commentId }} class="btn btn-sm text-danger font-weight-bold">
+                            <a href={{'index.php?admin=comment&delete=' ~ comment.id }} class="btn btn-sm text-danger font-weight-bold">
                                 <svg width="2em" height="2em" viewBox="0 0 16 16" class="bi bi-trash" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
                                     <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6z"/>
                                     <path fill-rule="evenodd" d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1zM4.118 4L4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118zM2.5 3V2h11v1h-11z"/>
@@ -58,7 +58,7 @@
                         </div>
                         <!--Valid-->
                         <div class="col-lg-6 ">
-                                <a href={{'index.php?p=post&id=' ~ comment.post_id ~ '&c=ok&cid=' ~ comment.commentId }} class="btn btn-sm text-blue font-weight-bold">
+                                <a href={{'index.php?p=post&id=' ~ comment.post_id ~ '&c=ok&cid=' ~ comment.id }} class="btn btn-sm text-blue font-weight-bold">
                                     <svg width="2em" height="2em" viewBox="0 0 16 16" class="bi bi-check2-circle" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
                                         <path fill-rule="evenodd" d="M15.354 2.646a.5.5 0 0 1 0 .708l-7 7a.5.5 0 0 1-.708 0l-3-3a.5.5 0 1 1 .708-.708L8 9.293l6.646-6.647a.5.5 0 0 1 .708 0z"/>
                                         <path fill-rule="evenodd" d="M8 2.5A5.5 5.5 0 1 0 13.5 8a.5.5 0 0 1 1 0 6.5 6.5 0 1 1-3.25-5.63.5.5 0 1 1-.5.865A5.472 5.472 0 0 0 8 2.5z"/>
