+++ view/frontView/homeView.twig
@@ -70,7 +70,7 @@
                                 <div class="col-lg-6"><p class="text-right text-black-50">{{ post.authorName }}</p></div>
                             </div>
                             <p class="text-black-50 text-justify mb-0">{{ post.content|truncate(155,true) }}</p>
-                            <p class="mt-2"><a href='index.php?p=post&amp;id={{ post.postId }}' >LIRE LA SUITE >></a></p>
+                            <p class="mt-2"><a href='index.php?p=post&amp;id={{ post.id }}' >LIRE LA SUITE >></a></p>
                         </div>
                     </div>
                 </div>
@@ -88,7 +88,7 @@
                                         <div class="col-lg-6"><p class="text-right text-white-50">{{ post.authorName }}</p></div>
                                     </div>
                                     <p class="mb-0 text-justify text-white-50">{{ post.content|truncate(155,true) }}</p>
-                                    <p class="mt-2"><a href='index.php?p=post&amp;id={{ post.postId }}' >LIRE LA SUITE >></a></p>
+                                    <p class="mt-2"><a href='index.php?p=post&amp;id={{ post.id }}' >LIRE LA SUITE >></a></p>
                                     <hr class="d-none d-lg-block mb-0 ml-0" />
                                 </div>
                             </div>
@@ -121,7 +121,7 @@
                             <div class="col"><input class="form-control flex-fill mr-0 mr-sm-2 mb-3 mb-sm-0 m-3" id="inputFirstName" name="inputFirstName" type="text" placeholder="Votre prénom..." required /></div>
                         </div>
                         <input class="form-control flex-fill mr-0 mr-sm-2 mb-3 mb-sm-0 m-3" id="inputEmail" name="inputEmail" type="email" placeholder="Votre adresse mail..." required />
-                        <textarea class="form-control flex-fill mr-0 mr-sm-2 mb-3 mb-sm-0 m-3" id="inputMessage" name="inputMessage" rows="10" placeholder="Votre message ici..." required ></textarea>
+                        <textarea class="form-control flex-fill mr-0 mr-sm-2 mb-3 mb-sm-0 m-3" id="inputMessage" name="inputMessage" rows="10" placeholder="Votre message ici..." required></textarea>
                         <div class="row"></div><button class="btn btn-primary m-3" type="submit" id="submitMessage" name="submitMessage">Envoyer</button></div>
                         <div class="row"></div><p class="mb-0 text-white-50">{{ msg }}</p></div>
                     </div>
