+++ view/frontView/listPostView.twig
@@ -19,7 +19,7 @@
                                 <div class="col-lg-6"><p class="text-right text-black-50">{{ post.authorName }}</p></div>
                             </div>
                             <p class="text-black-50 text-justify mb-0">{{ post.content|truncate(155,true) }}</p>
-                            <p class="mt-2"><a href='index.php?p=post&amp;id={{ post.postId }}' >LIRE LA SUITE >></a></p>
+                            <p class="mt-2"><a href='index.php?p=post&amp;id={{ post.id }}' >LIRE LA SUITE >></a></p>
                         </div>
                     </div>
                 </div>
